backend/requirements.txt
Flask~=3.0
flask-cors~=4.0
sqlmodel~=0.0.14
psycopg2-binary~=2.9
bcrypt~=4.1
alembic~=1.13
python-dotenv~=1.0
google-api-python-client~=2.0
google-auth-httplib2~=0.2
google-auth-oauthlib~=1.0
pytest~=7.4
